Seperate samples from the library
