+++ .idea/gradle.xml
@@ -4,14 +4,15 @@
   <component name="GradleSettings">
     <option name="linkedExternalProjectsSettings">
       <GradleProjectSettings>
-        <option name="distributionType" value="DEFAULT_WRAPPED" />
         <option name="externalProjectPath" value="$PROJECT_DIR$" />
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/android" />
-            <option value="$PROJECT_DIR$/common" />
-            <option value="$PROJECT_DIR$/desktop" />
+            <option value="$PROJECT_DIR$/richeditor" />
+            <option value="$PROJECT_DIR$/sample" />
+            <option value="$PROJECT_DIR$/sample/android" />
+            <option value="$PROJECT_DIR$/sample/common" />
+            <option value="$PROJECT_DIR$/sample/desktop" />
           </set>
         </option>
       </GradleProjectSettings>
