--- android/build.gradle.kts
@@ -1,41 +1,0 @@
-plugins {
-    id("org.jetbrains.compose")
-    id("com.android.application")
-    kotlin("android")
-}
-
-group "com.mocoding"
-version "1.0-SNAPSHOT"
-
-android {
-    namespace = "com.mocoding.richeditor.android"
-    compileSdk = (findProperty("android.compileSdk") as String).toInt()
-
-    defaultConfig {
-        minSdk = (findProperty("android.minSdk") as String).toInt()
-        targetSdk = (findProperty("android.targetSdk") as String).toInt()
-
-        applicationId = "com.mocoding.richeditor.android"
-        versionCode = 1
-        versionName = "1.0"
-    }
-
-    buildFeatures {
-        compose = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = libs.versions.compose.compiler.get()
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-    kotlinOptions {
-        jvmTarget = "11"
-    }
-    dependencies {
-        implementation(project(":common"))
-
-        implementation(libs.activity.compose)
-    }
-}
