--- common/src/commonMain/kotlin/com/mocoding/richeditor/common/ui/RichText.kt
@@ -1,31 +1,0 @@
-package com.mocoding.richeditor.common.ui
-
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import com.mocoding.richeditor.common.model.RichTextValue
-
-@Composable
-fun RichText(
-    value: RichTextValue,
-    modifier: Modifier = Modifier,
-) {
-    Text(
-        modifier = modifier,
-        text = value.visualTransformation.filter(value.annotatedString).text,
-    )
-}
-
-@Composable
-fun defaultRichTextStyle() = RichTextStyle(
-    textStyle = MaterialTheme.typography.body1,
-    textColor = MaterialTheme.colors.onPrimary,
-)
-
-data class RichTextStyle(
-    val textStyle: TextStyle,
-    val textColor: Color,
-)
