--- common/src/desktopMain/kotlin/com/mocoding/richeditor/common/DesktopApp.kt
@@ -1,11 +1,0 @@
-package com.mocoding.richeditor.common
-
-import androidx.compose.desktop.ui.tooling.preview.Preview
-import androidx.compose.runtime.Composable
-import com.mocoding.richeditor.common.App
-
-@Preview
-@Composable
-fun AppPreview() {
-    App()
-}
