+++ richeditor/build.gradle.kts
@@ -1,0 +1,60 @@
+plugins {
+    kotlin("multiplatform")
+    kotlin("plugin.serialization")
+    id("org.jetbrains.compose")
+    id("com.android.library")
+}
+
+group = "com.mocoding"
+version = "1.0-SNAPSHOT"
+
+kotlin {
+    android()
+    jvm("desktop") {
+        jvmToolchain(11)
+    }
+
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                api(compose.runtime)
+                api(compose.foundation)
+                api(compose.material)
+                api(compose.material3)
+                api(compose.materialIconsExtended)
+
+                implementation(libs.kotlinx.serialization.json)
+            }
+        }
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test"))
+            }
+        }
+        val androidMain by getting {
+            dependencies {
+                api("androidx.appcompat:appcompat:1.6.1")
+            }
+        }
+        val desktopMain by getting {
+            dependencies {
+                api(compose.preview)
+            }
+        }
+    }
+}
+
+android {
+    namespace = "com.mocoding.richeditor"
+    compileSdk = (findProperty("android.compileSdk") as String).toInt()
+
+    defaultConfig {
+        minSdk = (findProperty("android.minSdk") as String).toInt()
+        targetSdk = (findProperty("android.targetSdk") as String).toInt()
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_1_8
+        targetCompatibility = JavaVersion.VERSION_1_8
+    }
+}
