+++ sample/android/src/main/AndroidManifest.xml
