+++ sample/desktop/build.gradle.kts
@@ -1,0 +1,37 @@
+import org.jetbrains.compose.desktop.application.dsl.TargetFormat
+
+plugins {
+    kotlin("multiplatform")
+    id("org.jetbrains.compose")
+}
+
+group = "com.mocoding"
+version = "1.0-SNAPSHOT"
+
+
+kotlin {
+    jvm {
+        jvmToolchain(11)
+        withJava()
+    }
+    sourceSets {
+        val jvmMain by getting {
+            dependencies {
+                implementation(project(":richeditor"))
+                implementation(compose.desktop.currentOs)
+            }
+        }
+        val jvmTest by getting
+    }
+}
+
+compose.desktop {
+    application {
+        mainClass = "MainKt"
+        nativeDistributions {
+            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
+            packageName = "compose-richeditor"
+            packageVersion = "1.0.0"
+        }
+    }
+}
