+++ settings.gradle.kts
@@ -9,4 +9,9 @@ pluginManagement {
 
 rootProject.name = "compose-richeditor"
 
-include(":android", ":desktop", ":common")
+include(
+    ":sample:android",
+    ":sample:desktop",
+    ":sample:common",
+    ":richeditor"
+)
